Use unsigned long long to define lastn values to be grabbed

+++ src/sls.h
@@ -9,7 +9,7 @@
 using namespace std;
 namespace sls{
     bool append(const char *key, string data);
-    list<sls::Value> *lastn(const char *key, int num_entries);
+    list<sls::Value> *lastn(const char *key, unsigned long long num_entries);
     list<sls::Value> *all(const char *key);
     list<sls::Value> *intervalt(const char *key, unsigned long long start, unsigned long long end);
     string unwrap(sls::Value value);
